BOAC-1180, delete unused /api/teams; consolidate student-related api code
