+++ boac/api/student_controller.py
@@ -24,15 +24,46 @@ ENHANCEMENTS, OR MODIFICATIONS.
 """
 
 
-from boac.api.errors import BadRequestError, ForbiddenRequestError
+from boac.api.errors import BadRequestError, ForbiddenRequestError, ResourceNotFoundError
 from boac.api.util import add_alert_counts, can_current_user_view_dept, is_current_user_asc_affiliated
+from boac.externals import data_loch
+from boac.externals.cal1card_photo_api import get_cal1card_photo
 from boac.lib import util
 from boac.lib.berkeley import get_dept_codes
 from boac.lib.http import tolerant_jsonify
-from boac.merged.student import query_students, search_for_students
+from boac.merged import athletics
+from boac.merged.student import get_student_and_terms, get_student_query_scope, query_students, search_for_students
 from boac.models.alert import Alert
-from flask import current_app as app, request
+from flask import current_app as app, request, Response
 from flask_login import current_user, login_required
+
+
+@app.route('/api/majors/relevant')
+def relevant_majors():
+    majors = [row['major'] for row in data_loch.get_majors(get_student_query_scope())]
+    return tolerant_jsonify(majors)
+
+
+@app.route('/api/student/<uid>/analytics')
+@login_required
+def user_analytics(uid):
+    feed = get_student_and_terms(uid)
+    if not feed:
+        raise ResourceNotFoundError('Unknown student')
+    # CalCentral's Student Overview page is advisors' official information source for the student.
+    feed['studentProfileLink'] = f'https://calcentral.berkeley.edu/user/overview/{uid}'
+    return tolerant_jsonify(feed)
+
+
+@app.route('/api/student/<uid>/photo')
+@login_required
+def user_photo(uid):
+    photo = get_cal1card_photo(uid)
+    if photo:
+        return Response(photo, mimetype='image/jpeg')
+    else:
+        # Status is NO_DATA
+        return Response('', status=204)
 
 
 @app.route('/api/students', methods=['POST'])
@@ -114,6 +145,15 @@ def search_students():
         'students': students,
         'totalStudentCount': results['totalStudentCount'],
     })
+
+
+@app.route('/api/team_groups/all')
+@login_required
+def get_all_team_groups():
+    # TODO: Give unauthorized user a 404 without disrupting COE advisors on the filtered-cohort view.
+    _authorized = current_user.is_admin or is_current_user_asc_affiliated()
+    data = athletics.all_team_groups() if _authorized else []
+    return tolerant_jsonify(data)
 
 
 def _convert_asc_inactive_arg(is_inactive_asc):
