+++ boac/api/user_controller.py
@@ -26,17 +26,14 @@ ENHANCEMENTS, OR MODIFICATIONS.
 
 from boac.api import errors
 from boac.api.util import admin_required, authorized_users_api_feed, can_current_user_view_dept, decorate_cohort
-from boac.externals import data_loch
-from boac.externals.cal1card_photo_api import get_cal1card_photo
 from boac.lib.berkeley import get_dept_codes
 from boac.lib.http import tolerant_jsonify
 from boac.merged import calnet
-from boac.merged.student import get_student_and_terms, get_student_query_scope
 from boac.models.authorized_user import AuthorizedUser
 from boac.models.cohort_filter import CohortFilter
 from boac.models.curated_cohort import CuratedCohort
 from boac.models.university_dept import UniversityDept
-from flask import current_app as app, Response
+from flask import current_app as app
 from flask_login import current_user, login_required
 
 
@@ -105,31 +102,3 @@ def all_user_profiles_of_dept(dept_code):
             raise errors.ForbiddenRequestError(f'{current_user.uid} is not authorized to access {dept_code}')
     else:
         raise errors.ResourceNotFoundError('Unknown path')
-
-
-@app.route('/api/user/<uid>/analytics')
-@login_required
-def user_analytics(uid):
-    feed = get_student_and_terms(uid)
-    if not feed:
-        raise errors.ResourceNotFoundError('Unknown student')
-    # CalCentral's Student Overview page is advisors' official information source for the student.
-    feed['studentProfileLink'] = f'https://calcentral.berkeley.edu/user/overview/{uid}'
-    return tolerant_jsonify(feed)
-
-
-@app.route('/api/majors/relevant')
-def relevant_majors():
-    majors = [row['major'] for row in data_loch.get_majors(get_student_query_scope())]
-    return tolerant_jsonify(majors)
-
-
-@app.route('/api/user/<uid>/photo')
-@login_required
-def user_photo(uid):
-    photo = get_cal1card_photo(uid)
-    if photo:
-        return Response(photo, mimetype='image/jpeg')
-    else:
-        # Status is NO_DATA
-        return Response('', status=204)
