+++ boac/static/app/cohort/_filtered/cohort.html
@@ -146,7 +146,7 @@
             <div class="cohort-list-view-column-01">
               <img class="student-avatar student-avatar-large"
                    data-ng-class="{'img-blur': demoMode.blur}"
-                   data-ng-src="/api/user/{{student.uid}}/photo"
+                   data-ng-src="/api/student/{{student.uid}}/photo"
                    data-avatar-fallback>
             </div>
