+++ boac/static/app/course/course.html
@@ -97,7 +97,7 @@
             <div class="course-list-view-column-01" data-ng-if="!$first">
               <img class="student-avatar student-avatar-large"
                    data-ng-class="{'img-blur': demoMode.blur}"
-                   data-ng-src="/api/user/{{student.uid}}/photo"
+                   data-ng-src="/api/student/{{student.uid}}/photo"
                    data-avatar-fallback>
             </div>
