+++ boac/static/app/home/sortableAlertsTable.html
@@ -39,7 +39,7 @@
     <div class="group-summary-column group-summary-column-01">
       <img class="student-avatar student-avatar-small"
            data-ng-class="{'img-blur': demoMode.blur}"
-           data-ng-src="/api/user/{{student.uid}}/photo"
+           data-ng-src="/api/student/{{student.uid}}/photo"
            data-avatar-fallback>
     </div>
     <div class="group-summary-column group-summary-column-02">
