+++ boac/static/app/shared/matrix.html
@@ -24,7 +24,7 @@
         <div class="student-avatar-container">
           <img class="student-avatar"
                data-ng-class="{'img-blur': demoMode.blur}"
-               data-ng-src="/api/user/{{student.uid}}/photo"
+               data-ng-src="/api/student/{{student.uid}}/photo"
                data-avatar-fallback>
         </div>
         <div class="cohort-student-bio-container">
