+++ boac/static/app/student/student.html
@@ -21,7 +21,7 @@
             <div class="profile-avatar-container">
               <img class="profile-avatar"
                    data-ng-class="{'img-blur': demoMode.blur}"
-                   data-ng-src="/api/user/{{student.uid}}/photo"
+                   data-ng-src="/api/student/{{student.uid}}/photo"
                    data-avatar-fallback/>
             </div>
           </div>
