+++ boac/templates/index.html
@@ -40,7 +40,6 @@
     <script src="/static/app/bootstrap.js"></script>
     <script src="/static/app/routes.js"></script>
     <script src="/static/app/admin/adminController.js"></script>
-    <script src="/static/app/athletics/athleticsFactory.js"></script>
     <script src="/static/app/athletics/teamsController.js"></script>
     <script src="/static/app/cohort/curated/createController.js"></script>
     <script src="/static/app/cohort/curated/cohortController.js"></script>
